pyproject.toml
[tool.poetry]
name = "py-hpsu-monitor"
version = "0.4.0"
description = ""
authors = ["Felix Stürmer <stuermer@weltenwort.de>"]
include = ["default-config.toml"]

[tool.poetry.dependencies]
python = "^3.7"
python-can = "^3.3.4"
typer = "^0.4.0"
tomlkit = "^0.8.0"
typing_extensions = "^4.0.1"
pydantic = "^1.9.0"
asyncio-mqtt = "^0.11.1"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pylint = "^2.12.2"
black = "^21.12b0"
pytest-watch = "^4.2.0"
fabric = {version = "^2.6.0", optional = true}
pytest-asyncio = "^0.17.2"
mock = "^4.0.3"
flake8 = "^3.9.2"
isort = "^5.9.3"

[tool.poetry.extras]
deploy = ["fabric"]

[tool.poetry.scripts]
py-hpsu-monitor = "py_hpsu_monitor.cli:app"

[tool.isort]
profile = "black"

[tool.pyright]
typeCheckingMode = "strict"
reportMissingTypeStubs = "warning"
reportImportCycles = "warning"
useLibraryCodeForTypes = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
